# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CommonBeacon REST API",
    "version": "1.0.0",
    "description": "Implemented /api/v1 contract. Same-origin cookie sessions, CSRF on every mutation. Roles include MEMBER, MODERATOR, ADMINISTRATOR. Ordinary conversation writes reject archived boards; reporting/moderation remain available. Text is plain text, trimmed as documented and bounded in UTF-16 units (JSON Schema maxLength counts Unicode code points). All page offsets must fit a signed 32-bit integer. Business failures use ProblemDetail; infrastructure failures can have other response shapes. Health probes /api/health and /healthz are deployment endpoints, outside this contract."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/admin/data/jobs": {
      "get": {
        "operationId": "listCompanyTransferJobs",
        "summary": "listCompanyTransferJobs",
        "tags": [
          "transfers"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "description": "Company jobs require current ADMINISTRATOR status. Newest first; nextCursor is opaque. All responses use Cache-Control: no-store. Current database permissions and requester ownership are required.",
        "parameters": [
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "description": "Filter within this namespace; omitted retains existing mixed history behavior.",
            "schema": {
              "type": "string",
              "enum": [
                "COMPANY_EXPORT",
                "COMPANY_IMPORT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferJobPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Sign in required",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden or invalid CSRF/grant/ticket",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Job not found (also another requester or wrong namespace)",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Transfer storage unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/data/jobs/{id}": {
      "get": {
        "operationId": "getCompanyTransferJob",
        "summary": "getCompanyTransferJob",
        "tags": [
          "transfers"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "description": "Company jobs require current ADMINISTRATOR status. All responses use Cache-Control: no-store. Current database permissions and requester ownership are required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferJob"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Sign in required",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden or invalid CSRF/grant/ticket",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Job not found (also another requester or wrong namespace)",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Transfer storage unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/data/jobs/{id}/cancel": {
      "post": {
        "operationId": "cancelCompanyTransferJob",
        "summary": "cancelCompanyTransferJob",
        "tags": [
          "transfers"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "description": "Company jobs require current ADMINISTRATOR status. Matching cancellation retries replay for 24 hours; a reused key with different job/version conflicts. No recent-auth grant required. All responses use Cache-Control: no-store. Current database permissions and requester ownership are required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferJob"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Sign in required",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden or invalid CSRF/grant/ticket",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Job not found (also another requester or wrong namespace)",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version/state/idempotency conflict or download already active",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "JSON request exceeds 4096 bytes",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Transfer storage unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferCancellation"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/data/jobs/{id}/download-ticket": {
      "post": {
        "operationId": "issueCompanyTransferDownloadTicket",
        "summary": "issueCompanyTransferDownloadTicket",
        "tags": [
          "transfers"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "description": "Company jobs require current ADMINISTRATOR status. Consumes a DOWNLOAD grant. Ticket is single-use, session/job bound and expires after 60 seconds. All responses use Cache-Control: no-store. Current database permissions and requester ownership are required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferToken"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Sign in required",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden or invalid CSRF/grant/ticket",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Job not found (also another requester or wrong namespace)",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version/state/idempotency conflict or download already active",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "410": {
            "description": "Artifact unavailable or expired",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "JSON request exceeds 4096 bytes",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Transfer storage unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferTicketRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID",
        "description": "HttpOnly SameSite=Lax session cookie. Local HTTP profile disables Secure; production profile enables it. In-memory sessions expire on backend restart."
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ]
      },
...
      "TransferToken": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "token",
          "expiresAt"
        ],
        "additionalProperties": false
      },
      "TransferCancellation": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "minimum": 0,
            "format": "int64"
          }
        },
        "required": [
          "expectedVersion"
        ],
        "additionalProperties": false
      },
      "TransferTicketRequest": {
        "type": "object",
        "properties": {
          "recentAuthGrant": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          }
        },
        "required": [
          "recentAuthGrant"
        ],
        "additionalProperties": false
      },
      "TransferJob": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "COMPANY_EXPORT",
              "PERSONAL_EXPORT",
              "COMPANY_IMPORT"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "READY",
              "FAILED",
              "CANCELLED",
              "UPLOADING",
              "UPLOADED",
              "VALIDATING",
              "REVIEW_REQUIRED",
              "READY_TO_COMMIT",
              "COMMITTING",
              "COMPLETED"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "processedRecords": {
            "type": "integer",
            "minimum": 0
          },
          "errorCode": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              null,
              "WORK_FAILED",
              "AUTHORIZATION_REVOKED",
              "ATTEMPTS_EXHAUSTED",
              "JOB_EXPIRED",
              "ARTIFACT_MISSING",
              "ACTIVATION_RECONCILIATION_REQUIRED",
              "INVALID_SOURCE_DATA",
              "TRANSFER_LIMIT_EXCEEDED",
              "SNAPSHOT_TIMEOUT"
            ]
          },
          "artifactAvailable": {
            "type": "boolean"
          },
          "allowedActions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CANCEL",
                "DOWNLOAD"
              ]
            }
          },
          "provider": {
            "type": "string",
            "enum": [
              "NATIVE",
              "DISCOURSE"
            ],
            "default": "NATIVE",
            "description": "Immutable import provider. Discourse accepts only CommonBeacon bundle v1 from pinned Discourse 3.5.0."
          }
        },
        "required": [
          "id",
          "kind",
          "state",
          "version",
          "createdAt",
          "updatedAt",
          "expiresAt",
          "processedRecords",
          "errorCode",
          "artifactAvailable",
          "allowedActions"
        ],
        "additionalProperties": false
      },
      "TransferJobPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/TransferJob"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
    }
  }
}